requirements.txt
pyserial==3.5
python-dotenv==1.2.1
